refactor: extract useMemorySheetUI — 673 lines of memory sheet DOM builders and handlers
