+++ src/views/ChatView.vue
@@ -96,6 +96,7 @@ import { useChatMessageDisplay, restoreVisibleSwipeState } from '@/composables/c
 import { useContextBreakdown } from '@/composables/chat/useContextBreakdown.js';
 import { useMessageSelection } from '@/composables/chat/useMessageSelection.js';
 import { useChatSearch } from '@/composables/chat/useChatSearch.js';
+import { useMemorySheetUI } from '@/composables/chat/useMemorySheetUI.js';
 import { normalizeImgGenHtmlForEditing, prepareEditText, restoreEditText } from '@/core/utils/messageEditHelpers.js';
 
 function genMsgId() {
@@ -125,80 +126,6 @@ const chatGenerationServices = createChatGenerationServices({
     updateSessionMessage,
     runMemoryAutomationAfterStableTurn
 });
-
-async function openMemoryEntryEditor(entryId) {
-    if (!activeChatChar || !entryId) return;
-
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const entry = memoryBook.entries.find(item => item.id === entryId);
-    if (!entry) return;
-
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    content.innerHTML = `
-        <div class="settings-item">
-            <label>Title</label>
-            <input id="memory-entry-title" type="text" value="${(entry.title || '').replace(/"/g, '&quot;')}" placeholder="Memory title">
-        </div>
-        <div class="settings-item">
-            <label>Content</label>
-            <textarea id="memory-entry-content" rows="8" placeholder="Memory text">${(entry.content || '').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</textarea>
-        </div>
-        <div class="settings-item">
-            <label>Keys</label>
-            <input id="memory-entry-keys" type="text" value="${(Array.isArray(entry.keys) ? entry.keys.join(', ') : '').replace(/"/g, '&quot;')}" placeholder="key one, key two">
-            <div class="context-sheet-note">Only this field is used for keyword retrieval.</div>
-        </div>
-        <div class="context-sheet-actions">
-            <button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-entry-cancel">Cancel</button>
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-entry-save">Save</button>
-        </div>
-    `;
-
-    content.querySelector('#memory-entry-cancel')?.addEventListener('click', () => {
-        closeBottomSheet();
-        setTimeout(() => openMemoryTextPreview(entry, 'Memory Entry'), 50);
-    });
-
-    content.querySelector('#memory-entry-save')?.addEventListener('click', async () => {
-        const nextTitle = content.querySelector('#memory-entry-title')?.value?.trim() || 'Untitled memory';
-        const nextContent = content.querySelector('#memory-entry-content')?.value?.trim() || '';
-        const nextKeys = parseMemoryKeyInput(content.querySelector('#memory-entry-keys')?.value);
-
-        if (!nextContent) {
-            showToast('Memory content is required');
-            return;
-        }
-
-        const retrievalChanged = JSON.stringify(entry.keys || []) !== JSON.stringify(nextKeys)
-            || String(entry.content || '') !== nextContent;
-
-        entry.title = nextTitle;
-        entry.content = nextContent;
-        entry.keys = nextKeys;
-        entry.updatedAt = Date.now();
-        normalizeMemoryEntryShape(entry);
-        memoryBook.updatedAt = Date.now();
-        reconcileSessionMemoryState(chatData, sessionId, currentMessages.value);
-        chatData.sessions[sessionId] = currentMessages.value;
-
-        try {
-            if (getMemoryVectorSearchEnabled(memoryBook) && retrievalChanged) {
-                await reindexMemoryEntry(entry, activeChatChar.id, sessionId);
-            }
-            await db.saveChat(activeChatChar.id, chatData);
-            closeBottomSheet();
-            setTimeout(() => openMemoryTextPreview(entry, 'Memory Entry'), 50);
-        } catch (error) {
-            console.error('Failed to save memory entry:', error);
-            showToast(`Memory save failed: ${formatError(error)}`);
-        }
-    });
-
-    showBottomSheet({ title: 'Edit Memory Entry', content, isSolid: true });
-}
 
 const isAndroid = Capacitor.getPlatform() === 'android';
 const isBatterySaverUI = computed(() => shouldUseBatterySaverUI());
@@ -421,680 +348,54 @@ watch([isSearchMode, isSelectionMode], () => {
 });
 
 
-function openMemoryPromptPreview(item, options = {}) {
-    if (!item) return;
-    const { onClose } = options;
-    const hasOnClose = typeof onClose === 'function';
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    content.innerHTML = `
-        <div class="settings-item">
-            <label>Rule</label>
-            <div class="context-sheet-note">${(item.label || 'Prompt').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</div>
-        </div>
-        <div class="memory-entry-fulltext">${(item.prompt || '').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</div>
-        <div class="context-sheet-actions">
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-prompt-preview-close">${hasOnClose ? 'Back' : 'Close'}</button>
-        </div>
-    `;
-    content.querySelector('#memory-prompt-preview-close')?.addEventListener('click', () => {
-        closeBottomSheet();
-        if (hasOnClose) {
-            setTimeout(() => onClose(), 50);
-        }
-    });
-    showBottomSheet({ title: 'Generation Rule', content, isSolid: true });
-}
-
-async function createMemoryFromSelection() {
-    if (!activeChatChar || selectedMessages.value.size === 0) return;
-
-    const selected = currentMessages.value.filter(msg => msg && selectedMessages.value.has(msg.id));
-    if (!selected.length) return;
-
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const vectorEnabled = getMemoryVectorSearchEnabled(memoryBook);
-    const selectedIds = selected.map(msg => msg.id);
-    const firstMessage = selected[0];
-    const lastMessage = selected[selected.length - 1];
-    const previewLines = selected
-        .map(msg => `${msg.role === 'user' ? (msg.persona?.name || 'User') : (activeChatChar?.name || 'Character')}: ${msg.text || ''}`.trim())
-        .filter(Boolean)
-        .slice(0, 6);
-    const content = previewLines.join('\n').slice(0, 2000);
-    const personaNames = selected
-        .map(msg => msg.role === 'user' ? (msg.persona?.name || 'User') : (activeChatChar?.name || 'Character'))
-        .filter(Boolean);
-
-    const createdEntry = normalizeMemoryEntryShape({
-        id: genMemoryEntryId(),
-        title: `Memory ${memoryBook.entries.length + 1}`,
-        content,
-        keys: buildMemoryKeysFromText(content, personaNames),
-        glazeKeys: [],
-        vectorSearch: vectorEnabled,
-        messageIds: selectedIds,
-        messageRange: {
-            startMessageId: firstMessage.id,
-            endMessageId: lastMessage.id
-        },
-        status: 'active',
-        source: 'manual',
-        createdAt: Date.now(),
-        updatedAt: Date.now()
-    });
-    memoryBook.entries.push(createdEntry);
-
-    memoryBook.updatedAt = Date.now();
-    reconcileSessionMemoryState(chatData, sessionId, currentMessages.value);
-    chatData.sessions[sessionId] = currentMessages.value;
-    await db.saveChat(activeChatChar.id, chatData);
-    await indexMemoryEntryIfNeeded(createdEntry, activeChatChar.id, sessionId);
-    clearSelection();
-}
-
-async function generateMemoryDraftFromSelection() {
-    if (!activeChatChar || selectedMessages.value.size === 0) return;
-
-    const selected = currentMessages.value.filter(msg => msg && selectedMessages.value.has(msg.id));
-    if (!selected.length) return;
-    clearSelection();
-    await generateMemoryDraftForMessages(selected, { openSheet: true, source: 'manual_draft' });
-}
-
-function openMemoryTextPreview(entry, kind = 'Memory') {
-    if (!entry) return;
-    const keys = Array.isArray(entry.keys) && entry.keys.length
-        ? entry.keys.map(key => `<span class="memory-chip">${String(key).replace(/</g, '&lt;').replace(/>/g, '&gt;')}</span>`).join('')
-        : '<span class="context-sheet-note">No keys yet</span>';
-    const isApprovedEntry = kind === 'Memory Entry';
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    content.innerHTML = `
-        <div class="settings-item">
-            <label>${kind}</label>
-            <div class="context-sheet-note">${(entry.title || kind).replace(/</g, '&lt;').replace(/>/g, '&gt;')}</div>
-        </div>
-        <div class="settings-item">
-            <label>Retrieval</label>
-            <div class="context-sheet-note">Vector search: ${entry.vectorSearch ? 'enabled' : 'disabled'}</div>
-            <div class="memory-chip-list">${keys}</div>
-        </div>
-        <div class="memory-entry-fulltext">${(entry.content || '').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</div>
-        <div class="context-sheet-actions">
-            <button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-preview-regenerate">Regenerate</button>
-            ${isApprovedEntry ? `<button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-preview-edit">Edit</button>` : ''}
-            ${isApprovedEntry ? `<button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-preview-reindex">Reindex</button>` : ''}
-            ${isApprovedEntry ? `<button type="button" class="context-sheet-btn context-sheet-btn-secondary memory-preview-delete" id="memory-preview-delete">Delete</button>` : ''}
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-preview-close">${isApprovedEntry ? 'Close' : 'Back'}</button>
-        </div>
-    `;
-    content.querySelector('#memory-preview-edit')?.addEventListener('click', () => {
-        closeBottomSheet();
-        setTimeout(() => openMemoryEntryEditor(entry.id), 50);
-    });
-    content.querySelector('#memory-preview-reindex')?.addEventListener('click', async () => {
-        if (!activeChatChar) return;
-        const chatData = await getChatData(activeChatChar.id);
-        const sessionId = activeChatChar.sessionId || chatData.currentId;
-        const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-        if (!getMemoryVectorSearchEnabled(memoryBook)) {
-            showToast('Enable Memory Books vector search first');
-            return;
-        }
-        const reindexButton = content.querySelector('#memory-preview-reindex');
-        try {
-            reindexButton.disabled = true;
-            reindexButton.textContent = 'Reindexing...';
-            showToast('Reindexing memory entry...', 1500);
-            entry.vectorSearch = true;
-            await reindexMemoryEntry(entry, activeChatChar.id, sessionId);
-            entry.updatedAt = Date.now();
-            memoryBook.updatedAt = Date.now();
-            await db.saveChat(activeChatChar.id, chatData);
-            showToast('Memory entry reindexed');
-        } catch (error) {
-            console.error('Failed to reindex memory entry:', error);
-            showToast(`Reindex failed: ${formatError(error)}`);
-        } finally {
-            reindexButton.disabled = false;
-            reindexButton.textContent = 'Reindex';
-        }
-    });
-    content.querySelector('#memory-preview-delete')?.addEventListener('click', async () => {
-        if (!activeChatChar) return;
-        const chatData = await getChatData(activeChatChar.id);
-        const sessionId = activeChatChar.sessionId || chatData.currentId;
-        const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-        await deleteMemoryEntryIndexIfPresent(entry.id);
-        memoryBook.entries = memoryBook.entries.filter(item => item.id !== entry.id);
-        memoryBook.updatedAt = Date.now();
-        reconcileSessionMemoryState(chatData, sessionId, currentMessages.value);
-        chatData.sessions[sessionId] = currentMessages.value;
-        await db.saveChat(activeChatChar.id, chatData);
-        closeBottomSheet();
-        setTimeout(() => openMemoryBooksSheet(), 50);
-    });
-    content.querySelector('#memory-preview-regenerate')?.addEventListener('click', async () => {
-        if (!activeChatChar || !entry.messageIds || !entry.messageIds.length) {
-            showToast('Cannot regenerate: no message range');
-            return;
-        }
-        const chatData = await getChatData(activeChatChar.id);
-        const sessionId = activeChatChar.sessionId || chatData.currentId;
-        const messages = chatData.sessions[sessionId] || [];
-        const selectedMessages = messages.filter(msg => entry.messageIds.includes(msg.id));
-        if (!selectedMessages.length) {
-            showToast('Cannot regenerate: messages not found');
-            return;
-        }
-        closeBottomSheet();
-        // Regenerate draft
-        try {
-            await generateMemoryDraftForMessages(selectedMessages, { source: 'manual_regenerate' });
-            showToast('Draft regenerated');
-            setTimeout(() => openMemoryBooksSheet(), 50);
-        } catch (error) {
-            console.error('Failed to regenerate draft:', error);
-            showToast(`Regeneration failed: ${formatError(error)}`);
-        }
-    });
-    content.querySelector('#memory-preview-close')?.addEventListener('click', () => {
-        closeBottomSheet();
-        // Always return to Memory Books sheet
-        setTimeout(() => openMemoryBooksSheet(), 50);
-    });
-    showBottomSheet({ title: kind, content, isSolid: true });
-}
-
-async function openMessageMemoryCoverage(message) {
-    if (!activeChatChar || !message) return;
-
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const coverage = message.memoryCoverage && typeof message.memoryCoverage === 'object'
-        ? message.memoryCoverage
-        : createEmptyMemoryCoverage();
-    const entryIds = Array.isArray(coverage.entryIds) ? coverage.entryIds : [];
-    const matchedEntries = (Array.isArray(memoryBook.entries) ? memoryBook.entries : [])
-        .filter(entry => entryIds.includes(entry.id));
-
-    if (!matchedEntries.length) {
-        if (coverage.needsRebuild) {
-            showToast('This message is marked for memory rebuild');
-        } else if (coverage.stale) {
-            showToast('This message has stale memory coverage');
-        } else {
-            showToast('No linked memory entries for this message');
-        }
-        return;
-    }
-
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    content.innerHTML = `
-        <div class="settings-item">
-            <label>Message Memory Coverage</label>
-            <div class="context-sheet-note">This message is linked to ${matchedEntries.length} memory ${matchedEntries.length === 1 ? 'entry' : 'entries'}.</div>
-            ${coverage.needsRebuild ? '<div class="context-sheet-note" style="color: var(--warning-color, #ffb84d);">Coverage needs rebuild.</div>' : ''}
-            ${coverage.stale ? '<div class="context-sheet-note" style="color: var(--danger-color, #ff6b6b);">Coverage is marked stale.</div>' : ''}
-        </div>
-        <div class="memory-entry-list">
-            ${matchedEntries.map(entry => `
-                <button type="button" class="memory-entry-card" data-coverage-entry-id="${entry.id}">
-                    <div class="memory-entry-title-row">
-                        <strong>${String(entry.title || 'Memory Entry').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</strong>
-                        <span class="context-sheet-note">${String(entry.status || 'active').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</span>
-                    </div>
-                    <div class="context-sheet-note">${normalizeEntryMessageIds(entry).length} linked message${normalizeEntryMessageIds(entry).length === 1 ? '' : 's'}</div>
-                </button>
-            `).join('')}
-        </div>
-        <div class="context-sheet-actions">
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-coverage-close">Close</button>
-        </div>
-    `;
-
-    content.querySelectorAll('[data-coverage-entry-id]').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const entryId = btn.getAttribute('data-coverage-entry-id');
-            const entry = matchedEntries.find(item => item.id === entryId);
-            if (!entry) return;
-            closeBottomSheet();
-            setTimeout(() => openMemoryTextPreview(entry, 'Memory Entry'), 50);
-        });
-    });
-    content.querySelector('#memory-coverage-close')?.addEventListener('click', () => closeBottomSheet());
-    showBottomSheet({ title: 'Message Memory Coverage', content, isSolid: true });
-}
-
-async function removeMemoryFromSelection() {
-    if (!activeChatChar || selectedMessages.value.size === 0) return;
-
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const selectedIds = new Set(currentMessages.value.filter(msg => msg && selectedMessages.value.has(msg.id)).map(msg => msg.id));
-    if (!selectedIds.size) return;
-
-    const removedEntryIds = memoryBook.entries
-        .filter(entry => normalizeEntryMessageIds(entry).some(id => selectedIds.has(id)))
-        .map(entry => entry.id);
-
-    if (!removedEntryIds.length) {
-        clearSelection();
-        return;
-    }
-
-    memoryBook.entries = memoryBook.entries.filter(entry => !removedEntryIds.includes(entry.id));
-    memoryBook.updatedAt = Date.now();
-
-    for (const msg of currentMessages.value) {
-        if (!msg?.memoryCoverage) msg.memoryCoverage = createEmptyMemoryCoverage();
-        const wasCovered = Array.isArray(msg.memoryCoverage.entryIds) && msg.memoryCoverage.entryIds.some(id => removedEntryIds.includes(id));
-        msg.memoryCoverage.entryIds = (msg.memoryCoverage.entryIds || []).filter(id => !removedEntryIds.includes(id));
-        if (wasCovered) {
-            msg.memoryCoverage.needsRebuild = true;
-            msg.memoryCoverage.stale = false;
-        }
-    }
-
-    chatData.sessions[sessionId] = currentMessages.value;
-    await db.saveChat(activeChatChar.id, chatData);
-    clearSelection();
-}
-
-async function openMemoryGenerationSettings(initialState = null) {
-    if (!activeChatChar) return;
-
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    if (!memoryBook.settings) memoryBook.settings = {};
-    const currentApiConfig = getApiConfig();
-    const settings = memoryBook.settings;
-    const state = getNormalizedMemoryGenerationState(settings, initialState || {});
-
-    const renderSheet = () => {
-        const content = document.createElement('div');
-        content.className = 'context-sheet';
-        content.innerHTML = `
-            <div class="settings-item">
-                <label>Generation Rules</label>
-                <div class="clickable-selector" id="memory-prompt-selector">
-                    <span>${getMemoryPromptLabelByKey(settings, state.promptPreset)}</span>
-                    <svg viewBox="0 0 24 24"><path d="M7 10l5 5 5-5z"/></svg>
-                </div>
-                <button type="button" class="memory-inline-link" id="memory-prompt-preview-btn">Preview Rule</button>
-            </div>
-            <div class="settings-item">
-                <label>Temperature Override</label>
-                <input id="memory-temperature-input" type="number" min="0" max="2" step="0.05" value="${state.temperature ?? ''}" placeholder="Use current API temperature">
-            </div>
-            <div class="settings-item">
-                <label>Output Token Limit</label>
-                <input id="memory-max-tokens-input" type="number" min="200" max="32000" step="100" value="${state.maxTokens ?? ''}" placeholder="Auto (recommended 2000-4000 for large batches)">
-                <div class="context-sheet-note">Optional max completion tokens for memory draft generation. Leave blank to use the provider default with a safety floor.</div>
-            </div>
-            <div class="settings-item-checkbox">
-                <div class="settings-text-col">
-                    <label>Auto-Create Drafts</label>
-                    <div class="settings-desc">Automatically create Memory Book drafts after enough stable messages accumulate.</div>
-                </div>
-                <input id="memory-auto-create-toggle" type="checkbox" class="vk-switch" ${state.autoCreateEnabled ? 'checked' : ''}>
-            </div>
-            <div class="settings-item-checkbox">
-                <div class="settings-text-col">
-                    <label>Auto-Generate Draft Text</label>
-                    <div class="settings-desc">When enabled, newly auto-created draft placeholders immediately generate text. When disabled, auto mode only marks segments and leaves text generation manual.</div>
-                </div>
-                <input id="memory-auto-generate-toggle" type="checkbox" class="vk-switch" ${state.autoGenerateEnabled ? 'checked' : ''}>
-            </div>
-            <div class="settings-item">
-                <label>Create Memory Every N Messages</label>
-                <input id="memory-auto-interval-input" type="number" min="1" max="200" step="1" value="${state.autoCreateInterval}" placeholder="12">
-                <div class="context-sheet-note">User-facing interval for future automatic memory creation and import bootstrap segmentation.</div>
-            </div>
-            <div class="settings-item">
-                <label>Max Generate Batch</label>
-                <input id="memory-batch-size-input" type="number" min="1" max="50" step="1" value="${state.batchSize}" placeholder="1">
-                <div class="context-sheet-note">Limits how many pending drafts the batch generate button starts at once.</div>
-            </div>
-            <div class="settings-item-checkbox">
-                <div class="settings-text-col">
-                    <label>Work With Delay</label>
-                    <div class="settings-desc">Wait for extra turns before auto-creating a memory draft, so the last user message and latest assistant reply can still be edited or regenerated safely.</div>
-                </div>
-                <input id="memory-delayed-automation-toggle" type="checkbox" class="vk-switch" ${state.useDelayedAutomation ? 'checked' : ''}>
-            </div>
-            <div class="settings-item">
-                <label>Memory Entries In Prompt</label>
-                <input id="memory-max-injected-input" type="number" min="1" max="20" step="1" value="${state.maxInjectedEntries}" placeholder="3">
-                <div class="context-sheet-note">How many retrieved memory entries can be injected into the prompt at once.</div>
-            </div>
-            <div class="settings-item">
-                <label>Injection Target</label>
-                <div class="clickable-selector" id="memory-injection-target-selector">
-                    <span>${state.injectionTarget === 'summary_macro' ? '{{summary}} macro' : 'Chat summary block'}</span>
-                    <svg viewBox="0 0 24 24"><path d="M7 10l5 5 5-5z"/></svg>
-                </div>
-                <div class="context-sheet-note">Choose whether retrieved memory context follows the dedicated summary block path or the {{summary}} macro location.</div>
-            </div>
-            <div class="context-sheet-actions">
-                <button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-settings-cancel">Cancel</button>
-                <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-settings-save">Save</button>
-            </div>
-        `;
-
-        content.querySelector('#memory-prompt-selector')?.addEventListener('click', () => {
-            closeBottomSheet();
-            const promptItems = getMemoryPromptOptions(settings).map(item => ({
-                label: item.label,
-                onClick: () => {
-                    state.promptPreset = item.key;
-                    closeBottomSheet();
-                    setTimeout(() => showBottomSheet({ title: 'Memory Generation', content: renderSheet(), isSolid: true }), 50);
-                }
-            }));
-            promptItems.push({
-                label: `Preview: ${getMemoryPromptLabelByKey(settings, state.promptPreset)}`,
-                onClick: () => {
-                    const selected = getMemoryPromptOptions(settings).find(item => item.key === state.promptPreset);
-                    closeBottomSheet();
-                    setTimeout(() => openMemoryPromptPreview(selected, {
-                        onClose: () => openMemoryGenerationSettings({ ...state })
-                    }), 50);
-                }
-            });
-            promptItems.push({
-                label: 'Manage custom prompts',
-                onClick: () => {
-                    closeBottomSheet();
-                    setTimeout(() => openMemoryPromptManager(), 50);
-                }
-            });
-            showBottomSheet({ title: 'Generation Rules', items: promptItems });
-        });
-
-        content.querySelector('#memory-prompt-preview-btn')?.addEventListener('click', () => {
-            const options = getMemoryPromptOptions(settings);
-            const selected = options.find(item => item.key === state.promptPreset) || options[0];
-            closeBottomSheet();
-            setTimeout(() => openMemoryPromptPreview(selected, {
-                onClose: () => openMemoryGenerationSettings({ ...state })
-            }), 50);
-        });
-
-        content.querySelector('#memory-injection-target-selector')?.addEventListener('click', () => {
-            closeBottomSheet();
-            showBottomSheet({
-                title: 'Memory Injection Target',
-                items: [
-                    {
-                        label: 'Chat summary block',
-                        onClick: () => {
-                            state.injectionTarget = 'summary_block';
-                            closeBottomSheet();
-                            setTimeout(() => showBottomSheet({ title: 'Memory Generation', content: renderSheet(), isSolid: true }), 50);
-                        }
-                    },
-                    {
-                        label: '{{summary}} macro',
-                        onClick: () => {
-                            state.injectionTarget = 'summary_macro';
-                            closeBottomSheet();
-                            setTimeout(() => showBottomSheet({ title: 'Memory Generation', content: renderSheet(), isSolid: true }), 50);
-                        }
-                    }
-                ]
-            });
-        });
-
-        content.querySelector('#memory-settings-cancel')?.addEventListener('click', () => {
-            closeBottomSheet();
-            setTimeout(() => openMemoryBooksSheet(), 50);
-        });
-        content.querySelector('#memory-settings-save')?.addEventListener('click', async () => {
-            settings.generationSource = state.source;
-            settings.generationUseCurrentModelOverride = false;
-            settings.generationModel = state.model || '';
-            settings.generationEndpoint = state.source === 'custom'
-                ? (content.querySelector('#memory-endpoint-input')?.value?.trim() || '')
-                : '';
-            settings.generationApiKey = state.source === 'custom'
-                ? (content.querySelector('#memory-apikey-input')?.value || '')
-                : '';
-            const tempValue = content.querySelector('#memory-temperature-input')?.value?.trim();
-            settings.generationTemperature = tempValue === '' ? null : Number(tempValue);
-            const maxTokensValue = content.querySelector('#memory-max-tokens-input')?.value?.trim();
-            settings.generationMaxTokens = maxTokensValue === ''
-                ? null
-                : Math.max(200, Math.min(32000, Number.isFinite(Number(maxTokensValue)) ? Math.round(Number(maxTokensValue)) : 2000));
-            settings.autoCreateEnabled = !!content.querySelector('#memory-auto-create-toggle')?.checked;
-            settings.autoGenerateEnabled = !!content.querySelector('#memory-auto-generate-toggle')?.checked;
-            const autoIntervalValue = Number(content.querySelector('#memory-auto-interval-input')?.value || state.autoCreateInterval || 12);
-            settings.autoCreateInterval = Math.max(1, Math.min(200, Number.isFinite(autoIntervalValue) ? Math.round(autoIntervalValue) : 12));
-            const batchSizeValue = Number(content.querySelector('#memory-batch-size-input')?.value || state.batchSize || 1);
-            settings.batchSize = Math.max(1, Math.min(50, Number.isFinite(batchSizeValue) ? Math.round(batchSizeValue) : 1));
-            settings.useDelayedAutomation = !!content.querySelector('#memory-delayed-automation-toggle')?.checked;
-            const maxInjectedValue = Number(content.querySelector('#memory-max-injected-input')?.value || state.maxInjectedEntries || 3);
-            settings.maxInjectedEntries = Math.max(1, Math.min(20, Number.isFinite(maxInjectedValue) ? Math.round(maxInjectedValue) : 3));
-            settings.injectionTarget = state.injectionTarget === 'summary_macro' ? 'summary_macro' : 'summary_block';
-            settings.promptPreset = state.promptPreset || 'detailed_beats';
-            memoryBook.updatedAt = Date.now();
-            await db.saveChat(activeChatChar.id, chatData);
-            closeBottomSheet();
-            setTimeout(() => openMemoryBooksSheet(), 50);
-        });
-
-        return content;
-    };
-
-    showBottomSheet({
-        title: 'Memory Generation',
-        content: renderSheet(),
-        isSolid: true
-    });
-}
-
-async function openMemoryPromptManager() {
-    if (!activeChatChar) return;
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const settings = memoryBook.settings || {};
-    if (!Array.isArray(settings.customPrompts)) settings.customPrompts = [];
-
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    const promptCards = settings.customPrompts.length
-        ? settings.customPrompts.map(item => `
-            <div class="memory-entry-card" data-prompt-id="${item.id}">
-                <div class="memory-entry-head">
-                    <div>
-                        <div class="memory-entry-title">${(item.name || 'Custom prompt').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</div>
-                        <div class="memory-entry-meta">custom prompt</div>
-                    </div>
-                    <div class="memory-draft-actions">
-                        <button type="button" class="memory-entry-approve" data-prompt-edit="${item.id}">Edit</button>
-                        <button type="button" class="memory-entry-delete" data-prompt-delete="${item.id}">Delete</button>
-                    </div>
-                </div>
-                <div class="memory-entry-preview">${(item.prompt || '').replace(/</g, '&lt;').replace(/>/g, '&gt;').slice(0, 180)}</div>
-            </div>
-        `).join('')
-        : '<div class="context-sheet-note">No custom prompts yet.</div>';
-
-    content.innerHTML = `
-        <div class="context-sheet-actions" style="margin-top: 0; margin-bottom: 12px;">
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-prompt-add">Add Prompt</button>
-            <button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-prompt-close">Close</button>
-        </div>
-        <div class="memory-entry-list">${promptCards}</div>
-    `;
-
-    content.querySelector('#memory-prompt-add')?.addEventListener('click', () => {
-        closeBottomSheet();
-        setTimeout(() => openMemoryPromptEditor(), 50);
-    });
-    content.querySelector('#memory-prompt-close')?.addEventListener('click', () => closeBottomSheet());
-    content.querySelectorAll('[data-prompt-id]').forEach(card => {
-        card.addEventListener('click', (event) => {
-            if (event.target.closest('button')) return;
-            const promptId = card.getAttribute('data-prompt-id');
-            const prompt = settings.customPrompts.find(item => item.id === promptId);
-            if (!prompt) return;
-            closeBottomSheet();
-            setTimeout(() => openMemoryPromptPreview(
-                { label: prompt.name || 'Custom prompt', prompt: prompt.prompt || '' },
-                { onClose: openMemoryPromptManager }
-            ), 50);
-        });
-    });
-    content.querySelectorAll('[data-prompt-delete]').forEach(btn => {
-        btn.addEventListener('click', async () => {
-            const promptId = btn.getAttribute('data-prompt-delete');
-            settings.customPrompts = settings.customPrompts.filter(item => item.id !== promptId);
-            if (settings.promptPreset === promptId) settings.promptPreset = 'detailed_beats';
-            memoryBook.updatedAt = Date.now();
-            await db.saveChat(activeChatChar.id, chatData);
-            closeBottomSheet();
-            setTimeout(() => openMemoryPromptManager(), 50);
-        });
-    });
-    content.querySelectorAll('[data-prompt-edit]').forEach(btn => {
-        btn.addEventListener('click', () => {
-            const promptId = btn.getAttribute('data-prompt-edit');
-            const prompt = settings.customPrompts.find(item => item.id === promptId);
-            if (!prompt) return;
-            closeBottomSheet();
-            setTimeout(() => openMemoryPromptEditor(prompt), 50);
-        });
-    });
-
-    showBottomSheet({ title: 'Generation Rules', content, isSolid: true });
-}
-
-async function openMemoryPromptEditor(existing = null) {
-    if (!activeChatChar) return;
-    const chatData = await getChatData(activeChatChar.id);
-    const sessionId = activeChatChar.sessionId || chatData.currentId;
-    const memoryBook = ensureSessionMemoryBook(chatData, sessionId);
-    const settings = memoryBook.settings || {};
-    if (!Array.isArray(settings.customPrompts)) settings.customPrompts = [];
-
-    const content = document.createElement('div');
-    content.className = 'context-sheet';
-    content.innerHTML = `
-        <div class="settings-item">
-            <label>Name</label>
-            <input id="memory-prompt-name" type="text" value="${(existing?.name || '').replace(/"/g, '&quot;')}" placeholder="Prompt name">
-        </div>
-        <div class="settings-item">
-            <label>Prompt</label>
-            <textarea id="memory-prompt-text" rows="10" placeholder="Use {{history}}, {{user}}, {{char}}">${(existing?.prompt || '').replace(/</g, '&lt;').replace(/>/g, '&gt;')}</textarea>
-        </div>
-        <div class="context-sheet-actions">
-            <button type="button" class="context-sheet-btn context-sheet-btn-secondary" id="memory-prompt-cancel">Cancel</button>
-            <button type="button" class="context-sheet-btn context-sheet-btn-primary" id="memory-prompt-save">Save</button>
-        </div>
-    `;
-
-    content.querySelector('#memory-prompt-cancel')?.addEventListener('click', () => closeBottomSheet());
-    content.querySelector('#memory-prompt-save')?.addEventListener('click', async () => {
-        const name = content.querySelector('#memory-prompt-name')?.value?.trim() || 'Custom prompt';
-        const prompt = content.querySelector('#memory-prompt-text')?.value?.trim() || '';
-        if (!prompt) {
-            showToast('Prompt text is required');
-            return;
-        }
-        if (existing) {
-            const target = settings.customPrompts.find(item => item.id === existing.id);
-            if (target) {
-                target.name = name;
-                target.prompt = prompt;
-            }
-        } else {
-            const created = { id: genMemoryPromptId(), name, prompt };
-            settings.customPrompts.push(created);
-            settings.promptPreset = created.id;
-        }
-        memoryBook.updatedAt = Date.now();
-        await db.saveChat(activeChatChar.id, chatData);
-        closeBottomSheet();
-        setTimeout(() => openMemoryPromptManager(), 50);
-    });
-
-    showBottomSheet({ title: existing ? 'Edit Prompt' : 'Add Prompt', content, isSolid: true });
-}
-
-async function openMemoryBooksSheet() {
-    if (!activeChatChar) return;
-    await loadCurrentMemoryBook(activeChatChar);
-    nextTick(() => {
-        memoryBooksSheet.value?.open();
-    });
-}
-
-// Memory Books Sheet event handlers (wrappers for composable handlers)
-async function handleMemorySearchTypeUpdate() {
-    await handleMemorySearchTypeUpdate_composable(activeChatChar, memoryBooksSheet);
-}
-
-async function handleMemoryReindexAll() {
-    await handleMemoryReindexAll_composable(activeChatChar, memoryBooksSheet);
-}
-
-async function handleMemoryScanChat() {
-    await handleMemoryScanChat_composable(activeChatChar, currentMessages.value, memoryBooksSheet);
-}
-
-async function handleMemoryBatchGenerate() {
-    await handleMemoryBatchGenerate_impl();
-}
-
-async function handleMemoryGenerateSingleDraft(draftId) {
-    await generateSingleDraft(draftId);
-}
-
-async function handleMemoryApproveDraft(draftId) {
-    await handleMemoryApproveDraft_composable(draftId, activeChatChar, currentMessages.value, memoryBooksSheet);
-}
-
-async function handleMemoryDeleteDraft(draftId) {
-    await handleMemoryDeleteDraft_composable(draftId, activeChatChar, memoryBooksSheet);
-}
-
-async function handleMemoryDeleteAllDrafts() {
-    await handleMemoryDeleteAllDrafts_composable(activeChatChar, memoryBooksSheet);
-}
-
-async function handleMemoryDeleteEntry(entryId) {
-    await handleMemoryDeleteEntry_composable(entryId, activeChatChar, currentMessages.value, memoryBooksSheet);
-}
-
-function handleMemoryOpenMaintenance() {
-    handleMemoryOpenMaintenance_composable(activeChatChar, memoryBooksSheet);
-}
-
-function handleMemoryCancelDraft(draftId = null) {
-    cancelMemoryDraft(draftId || null);
-}
-
-function handleMemoryPreview({ entry, kind }) {
-    openMemoryTextPreview(entry, kind);
-}
-
-function handleMemoryOpenSettings() {
-    openMemoryGenerationSettings();
-}
-
-function handleMemoryQuickModelChange(model) {
-    handleMemoryQuickModelChange_impl(model);
-}
+const {
+    openMemoryEntryEditor,
+    openMemoryPromptPreview,
+    createMemoryFromSelection,
+    generateMemoryDraftFromSelection,
+    openMemoryTextPreview,
+    openMessageMemoryCoverage,
+    removeMemoryFromSelection,
+    openMemoryGenerationSettings,
+    openMemoryPromptManager,
+    openMemoryPromptEditor,
+    openMemoryBooksSheet,
+    handleMemorySearchTypeUpdate,
+    handleMemoryReindexAll,
+    handleMemoryScanChat,
+    handleMemoryBatchGenerate,
+    handleMemoryGenerateSingleDraft,
+    handleMemoryApproveDraft,
+    handleMemoryDeleteDraft,
+    handleMemoryDeleteAllDrafts,
+    handleMemoryDeleteEntry,
+    handleMemoryOpenMaintenance,
+    handleMemoryCancelDraft,
+    handleMemoryPreview,
+    handleMemoryOpenSettings,
+    handleMemoryQuickModelChange
+} = useMemorySheetUI({
+    getActiveChatChar: () => activeChatChar,
+    currentMessages,
+    selectedMessages,
+    clearSelection,
+    memoryBooksSheet,
+    loadCurrentMemoryBook,
+    generateMemoryDraftForMessages,
+    generateSingleDraft,
+    cancelMemoryDraft,
+    handleMemorySearchTypeUpdate_composable,
+    handleMemoryReindexAll_composable,
+    handleMemoryScanChat_composable,
+    handleMemoryApproveDraft_composable,
+    handleMemoryDeleteDraft_composable,
+    handleMemoryDeleteAllDrafts_composable,
+    handleMemoryDeleteEntry_composable,
+    handleMemoryOpenMaintenance_composable,
+    handleMemoryBatchGenerate_impl,
+    handleMemoryQuickModelChange_impl,
+    debouncedUpdateContextCutoff: () => debouncedUpdateContextCutoff()
+});
 
 // --- Display Logic (Separators) ---
 const displayMessages = computed(() => {
